Fixed cast to (int16) which should be cast to (int32), was preventing using long exposure times
git-svn-id: https://subversion.xor.aps.anl.gov/synApps/areaDetector/trunk@15267 dc6c5ff5-0b8b-c028-a01f-ffb33f00fc8b

+++ ADApp/pvcamSrc/pvCam.cpp
@@ -1425,7 +1425,7 @@ void pvCam::initializeDetector (void)
 
     //Exposure Time
     status |= getDoubleParam(ADAcquireTime,  &dValue);
-    int16Parm = (int16) (dValue * 1000);
+    int16Parm = (int32) (dValue * 1000);
     status |= getIntegerParam(PVCamTriggerMode, &iValue);
     int16Parm2 = iValue;
     printf ("binX: %d, binY: %d, minx: %d, miny: %d, sizex: %d, sizey: %d, triggerMode: %d, exposureTime: %d\n",
